requirement.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
colorama==0.4.4
cors==1.0.1
Django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-taggit==1.5.1
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
filelock==3.0.12
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.1
Markdown==3.3.4
Pillow==8.3.1
pycparser==2.20
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.1
tldextract==3.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
